+++ src/app/services/play.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { PlayService } from './play.service';
+
+describe('PlayService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: PlayService = TestBed.get(PlayService);
+    expect(service).toBeTruthy();
+  });
+});
